requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
supabase==2.10.0
requests==2.31.0
openai==2.12.0
pytest>=7.0.0